Executing pre-compile tasks... Loading Ant configuration... Running Ant tasks... Running 'before' tasks Checking sources Copying resources... [Annunaki] Checking dependencies… [Annunaki] Dependency analysis found 0 affected files Updating dependency information… [Annunaki] Parsing java… [Annunaki] Writing classes… [Annunaki] Dependency analysis found 29 affected files Adding @NotNull assertions… [Annunaki] Adding pattern assertions… [Annunaki] Adding the Threading Model assertions… [Annunaki] Dependency analysis found 0 affected files Building artifact 'Annunaki'… Deleting outdated files… Building artifact 'Annunaki': copying files… Building archives…
- Too many changes to list
- Removed constructors from checks
- Added alert delay
- Changed the way debug is handled
- Handle hit ticks in action processor
- Cleanup imports
- Added silent option

+++ src/main/java/fr/salers/annunaki/util/MathUtil.java
@@ -314,7 +314,7 @@ public class MathUtil {
         float f1 = MathHelper.sin(-yaw * 0.017453292F - (float) Math.PI);
         float f2 = -MathHelper.cos(-pitch * 0.017453292F);
         float f3 = MathHelper.sin(-pitch * 0.017453292F);
-        return new Vec3((double) (f1 * f2), (double) f3, (double) (f * f2));
+        return new Vec3(f1 * f2, f3, f * f2);
     }
 
     public static double[] getOffsetFromLocation(Location one, Location two) {
